silence signed/unsigned comparision warning

+++ lib/dns/dst_api.c
@@ -31,7 +31,7 @@
 
 /*
  * Principal Author: Brian Wellington
- * $Id: dst_api.c,v 1.11.92.8 2009/09/24 22:22:22 marka Exp $
+ * $Id: dst_api.c,v 1.11.92.9 2009/09/25 01:48:10 marka Exp $
  */
 
 /*! \file */
@@ -1248,7 +1248,7 @@ addsuffix(char *filename, unsigned int len, const char *ofilename,
 	n = snprintf(filename, len, "%.*s%s", olen, ofilename, suffix);
 	if (n < 0)
 		return (ISC_R_FAILURE);
-	if (n >= len)
+	if ((unsigned int)n >= len)
 		return (ISC_R_NOSPACE);
 	return (ISC_R_SUCCESS);
 }
